SSCS-10950 extracts helper methods and renames tests with @DisplayName
